Destructable Ship Modules and Projectiles
Ship modules can now be shot to the point where they disappear, and
allow projectiles to pass through them. Projectiles can now emanate from
inside a chunk, and are Armed once they pass outside of the ship hull.
When a ship hull is created, a duplicate sensor is overlaid on it, which
tells projectiles when to Arm(called on end contact).